change to single precision

+++ Constants.h
@@ -2,8 +2,8 @@
 
 #include <complex>
 
-constexpr double pi = 3.14159265358979;
-constexpr std::complex<double> i(0, 1);
+constexpr float pi = 3.14159265358979;
+constexpr std::complex<float> i(0, 1);
 
 enum class BoundaryCondition
 {
@@ -11,4 +11,4 @@ enum class BoundaryCondition
     Neumann
 };
 
-using complex = std::complex<double>;
+using complex = std::complex<float>;
