+++ Differentiation.h
@@ -7,8 +7,8 @@
 using namespace Eigen;
 
 
-MatrixXd VerticalDerivativeMatrix(BoundaryCondition originalBC, double L, int N);
-MatrixXd VerticalSecondDerivativeMatrix(BoundaryCondition bc, double L, int N);
+MatrixXf VerticalDerivativeMatrix(BoundaryCondition originalBC, float L, int N);
+MatrixXf VerticalSecondDerivativeMatrix(BoundaryCondition bc, float L, int N);
 
-DiagonalMatrix<double, -1> FourierSecondDerivativeMatrix(double L, int N, int dimension);
-DiagonalMatrix<complex, -1> FourierDerivativeMatrix(double L, int N, int dimension);
+DiagonalMatrix<float, -1> FourierSecondDerivativeMatrix(float L, int N, int dimension);
+DiagonalMatrix<complex, -1> FourierDerivativeMatrix(float L, int N, int dimension);
